+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 ### Breaking Changes
 
 - Move `prefect.client.result_handlers` to `prefect.engine.result_handlers` - [#512](https://github.com/PrefectHQ/prefect/pull/512)
+- Convert `timeout` kwarg from `timedelta` to `integer` - [#540](https://github.com/PrefectHQ/prefect/issues/540)
 
 ## 0.4.0 <Badge text="beta" type="success"/>